[ui] Clean up RunRow, move to a separate file
