+++ js_modules/dagster-ui/packages/ui-core/src/runs/RunTable.tsx
@@ -2,50 +2,23 @@ import {gql} from '@apollo/client';
 import {
   Box,
   Checkbox,
-  Colors,
   Icon,
   NonIdealState,
-  Table,
-  Mono,
-  Tag,
-  Button,
-  Dialog,
-  DialogBody,
-  DialogFooter,
-  ButtonLink,
   ProductTour,
   ProductTourPosition,
-  Caption,
+  Table,
 } from '@dagster-io/ui-components';
 import * as React from 'react';
-import {Link} from 'react-router-dom';
-import styled from 'styled-components';
 
-import {ShortcutHandler} from '../app/ShortcutHandler';
-import {isHiddenAssetGroupJob} from '../asset-graph/Utils';
 import {RunsFilter} from '../graphql/types';
 import {useSelectionReducer} from '../hooks/useSelectionReducer';
 import {useStateWithStorage} from '../hooks/useStateWithStorage';
-import {PipelineReference} from '../pipelines/PipelineReference';
 import {AnchorButton} from '../ui/AnchorButton';
-import {buildRepoAddress} from '../workspace/buildRepoAddress';
-import {useRepositoryForRunWithoutSnapshot} from '../workspace/useRepositoryForRun';
-import {workspacePipelinePath, workspacePipelinePathGuessRepo} from '../workspace/workspacePath';
 
-import {AssetKeyTagCollection, AssetCheckTagCollection} from './AssetTagCollections';
-import {CreatedByTagCell} from './CreatedByTag';
-import {RunActionsMenu, RunBulkActionsMenu} from './RunActionsMenu';
-import {RunStatusTagWithStats} from './RunStatusTag';
-import {DagsterTag, TagType} from './RunTag';
-import {RunTags} from './RunTags';
-import {
-  assetKeysForRun,
-  RunStateSummary,
-  RunTime,
-  RUN_TIME_FRAGMENT,
-  titleForRun,
-} from './RunUtils';
-import {RunFilterToken, runsPathWithFilters} from './RunsFilterInput';
+import {RunBulkActionsMenu} from './RunActionsMenu';
+import {RunRow} from './RunRow';
+import {RUN_TIME_FRAGMENT} from './RunUtils';
+import {RunFilterToken} from './RunsFilterInput';
 import ShowAndHideTagsMP4 from './ShowAndHideRunTags.mp4';
 import {RunTableRunFragment} from './types/RunTable.types';
 
@@ -249,293 +222,6 @@ export const RUN_TABLE_RUN_FRAGMENT = gql`
   ${RUN_TAGS_FRAGMENT}
 `;
 
-const RunRow = ({
-  run,
-  canTerminateOrDelete,
-  onAddTag,
-  checked,
-  onToggleChecked,
-  additionalColumns,
-  additionalActionsForRun,
-  isHighlighted,
-  hideCreatedBy,
-}: {
-  run: RunTableRunFragment;
-  canTerminateOrDelete: boolean;
-  onAddTag?: (token: RunFilterToken) => void;
-  checked?: boolean;
-  onToggleChecked?: (values: {checked: boolean; shiftKey: boolean}) => void;
-  additionalColumns?: React.ReactNode[];
-  additionalActionsForRun?: (run: RunTableRunFragment) => React.ReactNode[];
-  isHighlighted?: boolean;
-  hideCreatedBy?: boolean;
-}) => {
-  const {pipelineName} = run;
-  const repo = useRepositoryForRunWithoutSnapshot(run);
-
-  const isJob = React.useMemo(() => {
-    if (repo) {
-      const pipelinesAndJobs = repo.match.repository.pipelines;
-      const match = pipelinesAndJobs.find((pipelineOrJob) => pipelineOrJob.name === pipelineName);
-      return !!match?.isJob;
-    }
-    return false;
-  }, [repo, pipelineName]);
-
-  const repoAddressGuess = React.useMemo(() => {
-    if (repo) {
-      const {match} = repo;
-      return buildRepoAddress(match.repository.name, match.repositoryLocation.name);
-    }
-    return null;
-  }, [repo]);
-
-  const onChange = (e: React.FormEvent<HTMLInputElement>) => {
-    if (e.target instanceof HTMLInputElement) {
-      const {checked} = e.target;
-      const shiftKey =
-        e.nativeEvent instanceof MouseEvent && e.nativeEvent.getModifierState('Shift');
-      onToggleChecked && onToggleChecked({checked, shiftKey});
-    }
-  };
-
-  // Most system tags are unpinned by default so we track pinned for these.
-  const [pinnedSystemTags, setPinnedSystemTags] = useStateWithStorage(
-    'pinned-system-tags',
-    (value) => {
-      if (Array.isArray(value)) {
-        return value;
-      }
-      return [];
-    },
-  );
-
-  // All user tags are pinned by default so we track unpinned.
-  const [unpinnedTags, setUnpinnedTags] = useStateWithStorage('unpinned-tags', (value) => {
-    if (Array.isArray(value)) {
-      return value;
-    }
-    return [];
-  });
-
-  const allTagsWithPinned = React.useMemo(() => {
-    const allTags: Omit<(typeof run.tags)[0], '__typename'>[] = [...run.tags];
-    if ((isJob && run.mode !== 'default') || !isJob) {
-      allTags.push({
-        key: 'mode',
-        value: run.mode,
-      });
-    }
-    return allTags.map((tag) => {
-      return {
-        ...tag,
-        pinned: isUnpinnedByDefaultSystemTag(tag.key)
-          ? pinnedSystemTags.indexOf(tag.key) !== -1
-          : unpinnedTags.indexOf(tag.key) === -1,
-      };
-    });
-  }, [run, isJob, pinnedSystemTags, unpinnedTags]);
-
-  const isReexecution = run.tags.some((tag) => tag.key === DagsterTag.ParentRunId);
-
-  const targetBackfill = allTagsWithPinned.find((tag) => tag.key === DagsterTag.Backfill);
-
-  const [showRunTags, setShowRunTags] = React.useState(false);
-  const [isHovered, setIsHovered] = React.useState(false);
-
-  const tagsToShow = React.useMemo(() => {
-    const tagKeys: Set<string> = new Set([]);
-    const tags: TagType[] = [];
-    if (targetBackfill && targetBackfill.pinned) {
-      const link = run.assetSelection?.length
-        ? `/overview/backfills/${targetBackfill.value}`
-        : runsPathWithFilters([
-            {
-              token: 'tag',
-              value: `${DagsterTag.Backfill}=${targetBackfill.value}`,
-            },
-          ]);
-      tags.push({
-        ...targetBackfill,
-        link,
-      });
-      tagKeys.add(DagsterTag.Backfill as string);
-    }
-    allTagsWithPinned.forEach((tag) => {
-      if (tagKeys.has(tag.key)) {
-        // We already added this tag
-        return;
-      }
-      if (tag.pinned) {
-        tags.push(tag);
-      }
-    });
-    return tags;
-  }, [targetBackfill, allTagsWithPinned, run.assetSelection?.length]);
-
-  const onToggleTagPin = React.useCallback(
-    (tagKey: string) => {
-      if (isUnpinnedByDefaultSystemTag(tagKey)) {
-        setPinnedSystemTags((pinnedSystemTags) => toggleTag(pinnedSystemTags, tagKey));
-      } else {
-        setUnpinnedTags((unpinnedTags) => toggleTag(unpinnedTags, tagKey));
-      }
-    },
-    [setUnpinnedTags, setPinnedSystemTags],
-  );
-
-  return (
-    <Row
-      highlighted={!!isHighlighted}
-      onMouseEnter={() => {
-        setIsHovered(true);
-      }}
-      onMouseLeave={() => {
-        setIsHovered(false);
-      }}
-    >
-      {canTerminateOrDelete ? (
-        <td>{onToggleChecked ? <Checkbox checked={!!checked} onChange={onChange} /> : null}</td>
-      ) : null}
-      <td>
-        <Link to={`/runs/${run.id}`}>
-          <Mono>{titleForRun(run)}</Mono>
-        </Link>
-      </td>
-      <td>
-        <Box flex={{direction: 'column', gap: 4}}>
-          <RunTime run={run} />
-          {isReexecution ? (
-            <div>
-              <Tag icon="cached">Re-execution</Tag>
-            </div>
-          ) : null}
-        </Box>
-      </td>
-      <td style={{position: 'relative'}}>
-        <Box flex={{direction: 'column', gap: 5}}>
-          {isHiddenAssetGroupJob(run.pipelineName) ? (
-            <Box flex={{gap: 16}}>
-              <AssetKeyTagCollection assetKeys={assetKeysForRun(run)} />
-              <AssetCheckTagCollection assetChecks={run.assetCheckSelection} />
-            </Box>
-          ) : (
-            <Box flex={{direction: 'row', gap: 8, alignItems: 'center'}}>
-              <PipelineReference
-                isJob={isJob}
-                showIcon
-                pipelineName={run.pipelineName}
-                pipelineHrefContext="no-link"
-              />
-              <Link
-                to={
-                  repo
-                    ? workspacePipelinePath({
-                        repoName: repo.match.repository.name,
-                        repoLocation: repo.match.repositoryLocation.name,
-                        pipelineName: run.pipelineName,
-                        isJob,
-                      })
-                    : workspacePipelinePathGuessRepo(run.pipelineName)
-                }
-                target="_blank"
-              >
-                <Icon name="open_in_new" color={Colors.Blue500} />
-              </Link>
-            </Box>
-          )}
-          <Box
-            flex={{direction: 'row', alignItems: 'center', wrap: 'wrap'}}
-            style={{gap: '4px 8px'}}
-          >
-            <RunTagsWrapper>
-              {tagsToShow.length ? (
-                <RunTags tags={tagsToShow} onAddTag={onAddTag} onToggleTagPin={onToggleTagPin} />
-              ) : null}
-            </RunTagsWrapper>
-            {allTagsWithPinned.length > tagsToShow.length ? (
-              <Caption>
-                <ButtonLink
-                  onClick={() => {
-                    setShowRunTags(true);
-                  }}
-                  color={Colors.Gray700}
-                  style={{margin: '-4px', padding: '4px'}}
-                >
-                  View all tags ({allTagsWithPinned.length})
-                </ButtonLink>
-              </Caption>
-            ) : null}
-          </Box>
-        </Box>
-        {isHovered && allTagsWithPinned.length ? (
-          <ShortcutHandler
-            key="runtabletags"
-            onShortcut={() => {
-              setShowRunTags((showRunTags) => !showRunTags);
-            }}
-            shortcutLabel="t"
-            shortcutFilter={(e) => e.code === 'KeyT'}
-          >
-            {null}
-          </ShortcutHandler>
-        ) : null}
-      </td>
-      {hideCreatedBy ? null : (
-        <td>
-          <CreatedByTagCell
-            repoAddress={repoAddressGuess}
-            tags={run.tags || []}
-            onAddTag={onAddTag}
-          />
-        </td>
-      )}
-      <td>
-        <RunStatusTagWithStats status={run.status} runId={run.id} />
-      </td>
-      <td>
-        <RunStateSummary run={run} />
-      </td>
-      {additionalColumns}
-      <td>
-        <RunActionsMenu
-          run={run}
-          onAddTag={onAddTag}
-          additionalActionsForRun={additionalActionsForRun}
-        />
-      </td>
-      <Dialog
-        isOpen={showRunTags}
-        title="Tags"
-        canOutsideClickClose
-        canEscapeKeyClose
-        onClose={() => {
-          setShowRunTags(false);
-        }}
-      >
-        <DialogBody>
-          <RunTags tags={allTagsWithPinned} onAddTag={onAddTag} onToggleTagPin={onToggleTagPin} />
-        </DialogBody>
-        <DialogFooter topBorder>
-          <Button
-            intent="primary"
-            onClick={() => {
-              setShowRunTags(false);
-            }}
-          >
-            Close
-          </Button>
-        </DialogFooter>
-      </Dialog>
-    </Row>
-  );
-};
-
-const Row = styled.tr<{highlighted: boolean}>`
-  ${({highlighted}) =>
-    highlighted ? `box-shadow: inset 3px 3px #bfccd6, inset -3px -3px #bfccd6;` : null}
-`;
-
 function ActionBar({top, bottom}: {top: React.ReactNode; bottom?: React.ReactNode}) {
   return (
     <Box flex={{direction: 'column'}} padding={{vertical: 12}}>
@@ -554,31 +240,6 @@ function ActionBar({top, bottom}: {top: React.ReactNode; bottom?: React.ReactNod
       ) : null}
     </Box>
   );
-}
-
-const RunTagsWrapper = styled.div`
-  display: contents;
-  > * {
-    display: contents;
-  }
-`;
-
-function isUnpinnedByDefaultSystemTag(key: string) {
-  return (
-    (key.startsWith(DagsterTag.Namespace) &&
-      key !== DagsterTag.Partition &&
-      key !== DagsterTag.Backfill) ||
-    key === 'mode'
-  );
-}
-
-function toggleTag(tagsArr: string[] | undefined, tagKey: string): string[] {
-  const tags = tagsArr || [];
-  if (tags.indexOf(tagKey) !== -1) {
-    return tags.filter((key) => key !== tagKey);
-  } else {
-    return [...tags, tagKey];
-  }
 }
 
 function TargetHeader() {
